// src/types/distributor.ts
export type Distributor = {
  _id?: string
  name?: string
  photo?: any
  phone_no?: string | null
  email?: string | null
  address?: string | null
  note?: string
  to_received?: number
  last_received_amount?: number
  last_received?: number
  status?: string
  updatedAt?: string
  createdAt?: string
  deletedAt?: string
  claimed_amount?: number
  distributor_number?: number
}